docker環境の構築とreactのインストール
