+++ app/Providers/AppServiceProvider.php
@@ -23,7 +23,7 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        \URL::forceScheme('https');
+        \URL::forceScheme('http');
         $this->app['request']->server->set('HTTPS','on');
     }
 }
