+++ package.json
@@ -1,28 +1,30 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
+        "dev": "vite",
         "build": "vite build",
-        "watch": "mix watch",
-        "development": "mix",
         "heroku-postbuild": "npm run build"
     },
     "devDependencies": {
         "@tailwindcss/forms": "^0.5.2",
+        "@vitejs/plugin-react": "^4.2.1",
         "alpinejs": "^3.4.2",
         "autoprefixer": "^10.4.2",
         "axios": "^1.1.2",
-        "laravel-mix": "^6.0.49",
         "laravel-vite-plugin": "^0.7.2",
         "lodash": "^4.17.19",
         "postcss": "^8.4.6",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "resolve-url-loader": "^5.0.0",
         "sass": "^1.65.1",
         "sass-loader": "^12.6.0",
         "tailwindcss": "^3.1.0",
-        "vite": "^4.0.0"
+        "vite": "^4.0.0",
+        "vite-plugin-laravel": "^0.2.0-beta.9"
     },
     "dependencies": {
+        "@fortawesome/fontawesome-free": "^6.5.1",
         "font-awesome": "^4.7.0",
         "jquery": "^3.7.0",
         "laravel-echo": "^1.15.2",
