+++ resources/css/app.css
@@ -1,6 +1,7 @@
 @import 'tailwindcss/base';
 @import 'tailwindcss/components';
 @import 'tailwindcss/utilities';
+@import "@fortawesome/fontawesome-free/css/all.css";
 
 .liked{
     color:rgb(239, 68, 68);
