+++ resources/js/deal.jsx
