+++ resources/js/show.jsx
