+++ resources/scss/_variables.scss
