+++ resources/scss/app.scss
@@ -6,6 +6,3 @@
 
 // Bootstrap
 //@import 'bootstrap/scss/bootstrap';
-
-//font awesome
-@import "node_modules/font-awesome/scss/font-awesome.scss";
