+++ resources/views/components/header.blade.php
@@ -5,8 +5,9 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <meta name="csrf-token" content="{{ csrf_token() }}">
         <title>GoodsTrade</title>
-        <link href="/css/app.css" rel="stylesheet">
         <!-- Fonts -->
+        @viteReactRefresh
+        @vite(['resources/css/app.css', 'resources/js/app.jsx'])
     </head>
     <header class="border-b border-slate-200">
